requirements.txt
Flask==3.1.2
Flask-WTF==1.2.2
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3
Flask-Login==0.6.3
Flask-Migrate==4.1.0
flask-mysql-connector==1.1.0
PyMySQL==1.1.2
cryptography==46.0.3
python-dotenv==1.2.1
